agents/research_agent/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
python-dotenv>=1.0.0
tavily-python>=0.3.0
pydantic>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
